Refactor mergeInsertSort to use templates
Refactored the mergeInsertSort function and related sorting functions to use templates, allowing them to handle different container types generically. This improves code reusability and flexibility by eliminating duplicated code for lists and vectors and encapsulates the logic in a new PmergeMe.tpp file. Additionally, updated the main to use the templated version of mergeInsertSort for lists and vectors.